handle async email sending
