+++ backend/src/main/java/com/ecommerce/backend/shared/config/SpringAsyncConfig.java
@@ -1,0 +1,18 @@
+package com.ecommerce.backend.shared.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.task.SimpleAsyncTaskExecutor;
+import org.springframework.core.task.TaskExecutor;
+import org.springframework.scheduling.annotation.EnableAsync;
+
+import java.util.concurrent.Executor;
+
+@Configuration
+@EnableAsync
+public class SpringAsyncConfig {
+    @Bean
+    public TaskExecutor taskExecutor() {
+        return new SimpleAsyncTaskExecutor();
+    }
+}
