+++ backend/src/main/resources/application.yml
@@ -32,13 +32,14 @@ spring:
     properties:
       mail:
         smtp:
+          ssl:
+            trust: "*"
           auth: true
           connectiontimeout: 5000
           timeout: 5000
           writetimeout: 5000
           starttls:
             enable: true
-            required: true
   profiles:
     active: dev
